// data/github/current/ui/packages/issue-actions/__tests__/__generated__/IssueActionsTestQuery.graphql.ts
export type IssueActionsTestQuery$variables = {
  number: number;
  owner: string;
  repo: string;
};
export type IssueActionsTestQuery$data = {
  readonly repository: {
    readonly issue: {
      readonly " $fragmentSpreads": FragmentRefs<"IssueActions">;
    } | null | undefined;
  } | null | undefined;
};
export type IssueActionsTestQuery = {
  response: IssueActionsTestQuery$data;
  variables: IssueActionsTestQuery$variables;
};